backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
flask-sqlalchemy==3.0.3
flask-bcrypt==1.0.1
flask-jwt-extended==4.4.4
nltk==3.8.1
joblib==1.2.0
scikit-learn==1.2.2
pymysql==1.0.3

